metrics/requirements.txt
scikit-image
Pillow
numpy